// src/location/types/dadata.types.ts
export type AddressSuggestionsResponse = {
  suggestions: AddressSuggestion[];
};

export type AddressSuggestion = {
  value: string;
  data: AddressDetails;
};

export type AddressDetails = {
  area: string | null;
  area_fias_id: string | null;
  area_kladr_id: string | null;
  area_type: string | null;
  area_type_full: string | null;
  area_with_type: string | null;
  beltway_distance: string | null;
  beltway_hit: string | null;
  block: string | null;
  block_type: string | null;
  block_type_full: string | null;
  capital_marker: string;
  city: string | null;
  city_area: string | null;
  city_district: string | null;
  city_district_fias_id: string | null;
  city_district_kladr_id: string | null;
  city_district_type: string | null;
  city_district_type_full: string | null;
  city_district_with_type: string | null;
  city_fias_id: string;
  city_kladr_id: string;
  city_type: string;
  city_type_full: string;
  city_with_type: string;
  country: string;
  country_iso_code: string;
  divisions: string | null;
  entrance: string | null;
  federal_district: string;
  fias_actuality_state: string;
  fias_code: string | null;
  fias_id: string;
  fias_level: string;
  flat: string | null;
  flat_area: string | null;
  flat_cadnum: string | null;
  flat_fias_id: string | null;
  flat_price: string | null;
  flat_type: string | null;
  flat_type_full: string | null;
  floor: string | null;
  geo_lat: string | null;
  geo_lon: string | null;
  geoname_id: string;
  history_values: string | null;
  house: string | null;
  house_cadnum: string | null;
  house_fias_id: string | null;
  house_flat_count: string | null;
  house_kladr_id: string | null;
  house_type: string | null;
  house_type_full: string | null;
  kladr_id: string;
  metro: string | null;
  okato: string;
  oktmo: string;
  postal_box: string | null;
  postal_code: string;
  qc: string | null;
  qc_complete: string | null;
  qc_geo: string;
  qc_house: string | null;
  region: string;
  region_fias_id: string;
  region_iso_code: string;
  region_kladr_id: string;
  region_type: string;
  region_type_full: string;
  region_with_type: string;
  room: string | null;
  room_cadnum: string | null;
  room_fias_id: string | null;
  room_type: string | null;
  room_type_full: string | null;
  settlement: string | null;
  settlement_fias_id: string | null;
  settlement_kladr_id: string | null;
  settlement_type: string | null;
  settlement_type_full: string | null;
  settlement_with_type: string | null;
  source: string | null;
  square_meter_price: string | null;
  stead: string | null;
  stead_cadnum: string | null;
  stead_fias_id: string | null;
  stead_type: string | null;
  stead_type_full: string | null;
  street: string | null;
  street_fias_id: string | null;
  street_kladr_id: string | null;
  street_type: string | null;
  street_type_full: string | null;
  street_with_type: string | null;
  tax_office: string;
  tax_office_legal: string;
  timezone: string | null;
  unparsed_parts: string | null;
};